requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
google==3.0.0
newspaper3k==0.2.8
nltk==3.7
